% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnoutTools-package.R
\docType{package}
\name{burnoutTools}
\alias{burnoutTools-package}
\alias{burnoutTools}
\title{burnoutTools: Tools for Tracking and Preventing Burnout}
\description{
This package provides basic tools for simulating task-based burnout risk over time.
It combines simple probability models with visualization functions to help explore
how workload and recovery thresholds may affect long-term burnout risk.
}
\author{
\strong{Maintainer}: Tseeltuul Erdenebat \email{tseeleee.bay@gmail.com}

}
