% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnoutTools-package.R
\docType{package}
\name{burnoutTools}
\alias{burnoutTools-package}
\alias{burnoutTools}
\title{burnoutTools: Tools for tracking and preventing Burnout}
\description{
The burnoutTools package provides functions and models to help assess workload, rest patterns, and potential burnout risk.
}
\author{
\strong{Maintainer}: Tseeltuul Erdenebat \email{tseeleee.bay@gmail.com}

}
\keyword{internal}
