% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_burnout_forecast.R
\name{plot_burnout_forecast}
\alias{plot_burnout_forecast}
\title{Plot Burnout Forecast}
\usage{
plot_burnout_forecast(df)
}
\arguments{
\item{df}{Dataframe from simulate_burnout()}
}
\value{
A ggplot object
}
\description{
Plot Burnout Forecast
}
